+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.1.64](https://github.com/fohte/armyknife/compare/v0.1.63...v0.1.64) (2026-02-02)
+
+
+### Features
+
+* **gh/issue-agent:** manage title in frontmatter instead of body h1 ([#201](https://github.com/fohte/armyknife/issues/201)) ([c791583](https://github.com/fohte/armyknife/commit/c7915835908346c9839a7b94d955608374491562))
+
+## [0.1.63](https://github.com/fohte/armyknife/compare/v0.1.62...v0.1.63) (2026-02-02)
+
+
+### Bug Fixes
+
+* **gh/issue-agent:** correct misleading message after `init issue` ([#199](https://github.com/fohte/armyknife/issues/199)) ([fe35754](https://github.com/fohte/armyknife/commit/fe3575430fa521ffd394a25ce0101f5582718a3d))
+
 ## [0.1.62](https://github.com/fohte/armyknife/compare/v0.1.61...v0.1.62) (2026-02-01)